backend/pyproject.toml
[project]
name = "ledgerlens"
version = "2.0.0"
description = "Self-hosted, privacy-first receipt tracking"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.21",
    "alembic>=1.14",
    "pydantic-settings>=2.7",
    "itsdangerous>=2.2",
    "bcrypt>=4.2",
    "httpx>=0.28",
    "celery[redis]>=5.4",
    "rapidfuzz>=3.14",
    "pytesseract>=0.3",
    "pymupdf>=1.25",
    "pillow>=11.0",
    "openai>=1.60",
    "python-multipart>=0.0.18",
    "email-validator>=2.3.0",
]

[project.optional-dependencies]
postgres = ["asyncpg>=0.30"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.9",
    "mypy>=1.14",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=0.24",
]
